requirements.txt
beautifulsoup4==4.12.2
httpx==0.24.1
loguru==0.6.0
loguru==0.7.0
PyYAML==6.0.1
rich==13.4.2
